requirements.txt
Django==4.2.5
mysqlclient==2.2.0
Pillow==10.0.1
django-rest-framework==0.1.0
requests